app/requirements.txt
kafka-python
redis
psycopg2-binary  
docker
